Add parsing for RFC 3339 dates
These are mostly the same as ISO 8601 dates, but they are what chrono
uses. Now we attempt to parse a "human readable" date and then fall back
to RFC 3339 date parsing if that fails.

Note that I had to update chrono to 0.4.26 so that I could use the
`parse_and_remainder` function.